# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ITBF API",
        "description": "API for ITBF",
        "contact": {
            "name": "Christian Lerma",
            "url": "",
            "email": "christianlerma@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/habitaciones": {
            "get": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Get all registers",
                "operationId": "c7e7107182c27b4c083de81a8292ad4a",
                "responses": {
                    "200": {
                        "description": "Retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Create a new register",
                "operationId": "0e05066983a5258949c17768626221cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "habitacion",
                                    "descripcion",
                                    "hotel_id",
                                    "acomodacion_id",
                                    "tipo_id"
                                ],
                                "properties": {
                                    "habitacion": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "hotel_id": {
                                        "type": "integer"
                                    },
                                    "acomodacion_id": {
                                        "type": "integer"
                                    },
                                    "tipo_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/habitaciones/{id}": {
            "get": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Get a register by id",
                "operationId": "821a2af23dce2ec909b130196f1b52d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the register",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Habitacion not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/habitaciones/hotel/{id}": {
            "get": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Get all registers by hotel",
                "operationId": "d78916c9aae657b89df0a32dd58e6096",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/habitaciones/{id}/editar": {
            "put": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Update a register",
                "operationId": "44f69e23279380c787ebc63cab9af0de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the register",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "habitacion",
                                    "descripcion",
                                    "hotel_id",
                                    "acomodacion_id",
                                    "tipo_id"
                                ],
                                "properties": {
                                    "habitacion": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "hotel_id": {
                                        "type": "integer"
                                    },
                                    "acomodacion_id": {
                                        "type": "integer"
                                    },
                                    "tipo_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error updating the register",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/habitaciones/{id}/eliminar": {
            "delete": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Delete a register",
                "operationId": "91a9cc5bfb3b3c56ca3f48fe173cdc39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the register",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}